// src/models/IReceiving.ts
export interface IReceiving {
  id?: string;
  _id?: string;
  receivingNumber?: number;
  createUserID?: string;
  updateUserID?: string;
  createDate: Date;
  updateDate?: Date;
  receivingDate: Date;
  state?: TReceivingStatus;
  companyID?: string;
  receivingParts?: any[];
  orderItems?: any[];
  docs?: any[];
  supplier?: object;
  shipTo?: object;
  awbNumber?: string;
  awbType?: string;
  awbDate?: Date;
  awbReference?: string;
  AWB_NUMBER?: string;
  AWB_REFERENCE?: string;
  AWB_TYPE?: string;
  WAREHOUSE_CODE?: string;
  WAREHOUSE_AGREEMENT?: string;
  TRANSACTION_DATE?: any;
  STOCK?: any;
  CURRENCY?: any;
  NAME_OF_MATERIAL?: any;
  NAME_OF_MATERIAL_ENG?: any;
  WAREHOUSE_RECEIVED_AT?: any;
  SUPPLIES_CODE?: any;
  SUPPLIES_LOCATION?: any;
  SUPPLIER_SHORT_NAME?: any;
}
export type TReceivingStatus =
  | 'ARRIVED'
  | 'CLOSED'
  | 'MISSING'
  | 'OPEN'
  | 'OPEN_AND_TRANSFER'
  | 'PARTLY_ARRIVED'
  | 'PARTLY_MISSING'
  | 'PARTLY_SENT'
  | 'PARTLY_RECEIVED'
  | 'READY'
  | 'PARTLY_READY'
  | 'SENT'
  | 'TRANSFER'
  | 'UNKNOWN'
  | 'RECEIVED'
  | 'CANCELLED'
  | 'RECEIVED';

interface ReceivingPartData {
  RECEIVING_ID?: string;
  RECEIVING_NUMBER?: string;
  AWB_NUMBER?: string;
  AWB_REFERENCE?: string;
  AWB_TYPE?: string;
  AWB_DATE?: Date;
  TRANSACTION_DATE: Date;
  STOCK?: string;
  WAREHOUSE_CODE?: string;
  WAREHOUSE_AGREEMENT?: string;
  CURRENCY?: string;
  NAME_OF_MATERIAL?: string;
  NAME_OF_MATERIAL_ENG?: string;
  UNIT_OF_MEASURE?: string;
  UNIT_OF_MEASURE_ADD?: string;
  QUANTITY?: number;
  QUANTITY_ADD?: number;
  PRICE: number;
  PART_NUMBER?: string;
  RACK_NUMBER?: string;
  SHELF_NUMBER?: string;
  PRODUCT_EXPIRATION_DATE?: Date;
  INVENTORY_ACCOUNT?: string;
  CUSTOMER_MATERIAL?: boolean;
  OWNER_SHORT_NAME?: string;
  OWNER_LONG_NAME?: string;
  BATCH?: string;
  LOCAL_ID?: string;
  SUPPLIER_BATCH_NUMBER?: any;
  CONDITION: any;
  APPROVED_CERT?: any;
  RECEIVED_DATE?: Date;
}